+++ packages/polaris-common/src/notification-center/notification-center-config.ts
@@ -1,0 +1,10 @@
+import * as kafka from 'kafkajs';
+export interface NotificationCenterConfig {
+    allowNotificationCenter: boolean;
+    topicName: string;
+    kafkaClient: kafka.Kafka;
+    topicPrefix?: string;
+    topicsAmountOfPartition?: number;
+    partitionerSelector?: kafka.ICustomPartitioner;
+    topicsReplicationFactor?: number;
+}