+++ packages/polaris-core/src/config/polaris-server-config.ts
@@ -1,4 +1,4 @@
-import { RealitiesHolder } from '@enigmatis/polaris-common';
+import { NotificationCenterConfig, RealitiesHolder } from '@enigmatis/polaris-common';
 import { PolarisGraphQLLogger } from '@enigmatis/polaris-graphql-logger';
 import { ApplicationProperties } from '@enigmatis/polaris-logs';
 import { ConnectionlessConfiguration } from '@enigmatis/polaris-middlewares';
@@ -34,4 +34,5 @@ export interface PolarisServerConfig extends Omit<ApolloServerExpressConfig, 'lo
     permissionsConfig: PermissionsConfiguration;
     enableDataVersionFilter: boolean;
     connectionlessConfiguration?: ConnectionlessConfiguration;
+    notificationCenterConfig?: NotificationCenterConfig;
 }