+++ packages/polaris-core/src/index.ts
@@ -21,7 +21,11 @@ export { PermissionsConfiguration } from './config/permissions-configuration';
 export * from 'apollo-server-errors';
 export { PubSub, gql } from 'apollo-server-express';
 export * from '@enigmatis/polaris-common';
-export { Reality, ApplicationProperties } from '@enigmatis/polaris-common';
+export {
+    Reality,
+    ApplicationProperties,
+    NotificationCenterConfig,
+} from '@enigmatis/polaris-common';
 export * from '@enigmatis/polaris-logs';
 export * from '@enigmatis/polaris-graphql-logger';
 export * from '@enigmatis/polaris-middlewares';