+++ packages/polaris-permissions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/Enigmatis/polaris-permissions/compare/@enigmatis/polaris-permissions@1.5.0...@enigmatis/polaris-permissions@1.5.1) (2021-02-21)
+
+**Note:** Version bump only for package @enigmatis/polaris-permissions
+
+
+
+
+
 # [1.5.0](https://github.com/Enigmatis/polaris-permissions/compare/@enigmatis/polaris-permissions@1.2.6...@enigmatis/polaris-permissions@1.5.0) (2021-02-07)