+++ packages/polaris-schema/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/Enigmatis/polaris-schema/compare/@enigmatis/polaris-schema@1.7.0...@enigmatis/polaris-schema@1.7.1) (2021-02-21)
+
+**Note:** Version bump only for package @enigmatis/polaris-schema
+
+
+
+
+
 # [1.7.0](https://github.com/Enigmatis/polaris-schema/compare/@enigmatis/polaris-schema@1.5.2...@enigmatis/polaris-schema@1.7.0) (2021-02-07)